+++ app/gui/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 #### Visual Environment
 
+- [New nodes can be created by dragging and dropping a connection on the
+  scene.][3231]
+- [Node connections can be dropped by pressing the Esc key while dragging
+  them.][3231]
 - [Added support of source maps for JS-based visualizations.][3208]
 - [Fixed histograms coloring and added a color legend.][3153]
 - [Fixed broken node whose expression contains non-ASCII characters.][3166]
@@ -32,6 +36,7 @@
 [3208]: https://github.com/enso-org/enso/pull/3208
 [3224]: https://github.com/enso-org/enso/pull/3224
 [3229]: https://github.com/enso-org/enso/pull/3229
+[3231]: https://github.com/enso-org/enso/pull/3231
 [3232]: https://github.com/enso-org/enso/pull/3232
 
 # Enso 2.0.0-alpha.18 (2021-10-12)